changed defualt segment length to 1s, added code to convert mfa to wav, added code to resample files that arent 48000 to make them all the same length

+++ src/scripts/augment_segment_mfcc_extract_master.py
@@ -24,7 +24,9 @@ from joblib import Parallel, delayed
 
 #Saving files 
 import pickle
-#NOTE: YOU NEED TO pip install joblib librosa PySoundFile scipy pydub tqdm numpy pandas
+
+#NOTE: you need to pip install joblib librosa PySoundFile scipy pydub tqdm numpy pandas
+
  
 #This file takes multiple input directorys that are strutrcued as folder of car brands
 #This file augments the audio, segments all the audio, creates a dataframe out of the segments, and extracts mfccs and saves them for the top 10 brands and all brands
@@ -34,7 +36,7 @@ import pickle
 
 # Global variables
 SAMPLE_RATE = 48000
-SEGMENT_LENGTH_MS = 5000  # This can be overridden by command-line arguments
+SEGMENT_LENGTH_MS = 1000  # This can be overridden by command-line arguments
 N_MFCC = 13
 N_FFT = 2048
 HOP_LENGTH = 512
@@ -169,24 +171,24 @@ def lowpass_filter(audio_data, sample_rate):
 functions = [random_gain, noise_addition, pitch_shifting, highpass_filter, lowpass_filter]
 
 # Process file function
-def convert_m4a_to_wav(file_path):
+def convert_m4a_to_wav(file_path, no_convert=False):  
     """
     Converts an .m4a file to .wav format with the same base filename, then deletes the original .m4a file.
     Returns the new file path and conversion status.
     """
-    if file_path.endswith('.m4a'):
+    if file_path.endswith('.m4a') and not no_convert:  # Check the no_convert flag
         wav_file_path = file_path.rsplit('.', 1)[0] + '.wav'
         try:
-            ffmpeg.input(file_path).output(wav_file_path).run(overwrite_output=True , quiet=True)
+            ffmpeg.input(file_path).output(wav_file_path).run(overwrite_output=True, quiet=True)
             os.remove(file_path)  # Remove the original .m4a file
             return wav_file_path, "Converted"
         except ffmpeg.Error as e:
             return None, f"Conversion failed: {e}"
     return file_path, "Not converted"
 
-def process_and_segment_file(file_path, output_dir, functions):
+def process_and_segment_file(file_path, output_dir, functions, no_convert):
     # Before opening a file, check if it's .m4a and then convert
-    file_path, conversion_status = convert_m4a_to_wav(file_path)
+    file_path, conversion_status = convert_m4a_to_wav(file_path, no_convert)
 
     filename = os.path.basename(file_path)
     base_filename, _ = os.path.splitext(filename)
@@ -239,7 +241,7 @@ def process_and_segment_file(file_path, output_dir, functions):
     # Adding conversion_status to the return dict for additional feedback
     return {"path": file_path, "status": "Processed", "error": None, "conversion_status": conversion_status}
 
-def process_directories(input_dirs, output_dir, functions):
+def process_directories(input_dirs, output_dir, functions, no_content):
     print(f"Starting segmentation process for files in {input_dirs} to {output_dir} with segment length of {SEGMENT_LENGTH_MS}ms...\n")
     # Initialize a dictionary to hold the count of files per directory
     files_per_dir = {input_dir: 0 for input_dir in input_dirs}
@@ -270,19 +272,19 @@ def process_directories(input_dirs, output_dir, functions):
     failed_files = []
     
     with ThreadPoolExecutor(max_workers=os.cpu_count()) as executor:
-        futures = {executor.submit(process_and_segment_file, audio_path, output_dir, functions): audio_path for audio_path in audio_paths}
+        futures = {executor.submit(process_and_segment_file, audio_path, output_dir, functions, no_content): audio_path for audio_path in audio_paths}
         for future in tqdm(as_completed(futures), total=len(futures), desc="Processing"):
             audio_path = futures[future]
             result = future.result()
             if result["status"] == "Processed":
                 processed_files.append(audio_path)
                 # Check for conversion status
-                if result.get("conversion_status") == "Converted":
-                    converted_files.append(audio_path)
             elif result["status"] == "Skipped":
                 skipped_files.append(audio_path)
             elif result["status"] == "Error":
                 failed_files.append((audio_path, result["error"]))
+            if result.get("conversion_status") == "Converted":
+                converted_files.append(audio_path)
 
     print(f"\n\nOriginal files processed: {len(processed_files)}")
     print(f"Original files skipped: {len(skipped_files)}")
@@ -292,10 +294,10 @@ def process_directories(input_dirs, output_dir, functions):
         for file, error in failed_files:
             print(f"{file}: {error}")
 
-    # if processed_files:
-    #     print("process the following files:")
-    #     for file in processed_files:
-    #         print(f"{file}:")
+    if processed_files:
+        print("process the following files:")
+        for file in processed_files:
+            print(f"{file}:")
 
     # Reporting on converted files
     if converted_files:
@@ -349,11 +351,21 @@ def process_files_to_dataframe(path, output_dir):
 
 #MFCC extraction
 #############################################################################################
-def extract_features(audio_path, sample_rate, n_mfcc, n_fft, hop_length, expected_num_mfcc_vectors):
+def extract_features(audio_path, sample_rate, target_sample_rate, n_mfcc, n_fft, hop_length, expected_num_mfcc_vectors):
+    # Load the audio
     signal, sr = librosa.load(audio_path, sr=sample_rate)
+    
+    # Resample the audio if the sample rate is different from the target sample rate
+    if sr != target_sample_rate:
+        print(f"resample file: {audio_path}")
+        signal = librosa.resample(signal, orig_sr=sr, target_sr=target_sample_rate)
+        sr = target_sample_rate
+
+    # Extract MFCC features
     mfcc = librosa.feature.mfcc(y=signal, sr=sr, n_mfcc=n_mfcc, n_fft=n_fft, hop_length=hop_length)
     mfcc = mfcc.T
 
+    # Adjust the number of MFCC vectors
     if len(mfcc) > expected_num_mfcc_vectors:
         mfcc = mfcc[:expected_num_mfcc_vectors]
     elif len(mfcc) < expected_num_mfcc_vectors:
@@ -451,7 +463,7 @@ def save_mapping(folder_to_class_number, output_dir, filename):
 
 
 
-def main(onedrive_enabaled, augmentation_enabled, dataframe_creation_enabled, feature_extraction_enabled, input_dirs, output_dir, mfcc_filename,mfcc_output_dir,segment_length_ms, n_mfcc, set_brands):
+def main(onedrive_enabaled, augmentation_enabled, dataframe_creation_enabled, feature_extraction_enabled, input_dirs, output_dir, mfcc_filename,mfcc_output_dir,segment_length_ms, n_mfcc, set_brands, no_convert):
     global N_MFCC, SEGMENT_LENGTH_MS, DURATION, EXPECTED_NUM_MFCC_VECTORS
     N_MFCC=n_mfcc
     SEGMENT_LENGTH_MS = segment_length_ms
@@ -505,10 +517,10 @@ def main(onedrive_enabaled, augmentation_enabled, dataframe_creation_enabled, fe
         # Define your directories and call the function...
         functions = [random_gain, noise_addition, pitch_shifting, highpass_filter, lowpass_filter]
 
-        processed_files = process_directories(input_dirs, output_dir, functions)
+        processed_files = process_directories(input_dirs, output_dir, functions, no_convert)
     else:
         print("\n\n\nSkipping audio augmentation\nBeginning audio segmentation")
-        processed_files = process_directories(input_dirs, output_dir, [])
+        processed_files = process_directories(input_dirs, output_dir, [], no_convert)
     
 
 
@@ -570,9 +582,10 @@ if __name__ == "__main__":
     parser.add_argument('--output_dir', help='Output directory for segments (Will create a DIR containing folders of the same brand names from input_dir)(Use absolute paths)', default=None)
     parser.add_argument('--mfcc_filename', help='Change the file name of the mfcc output', default=None)
     parser.add_argument('--mfcc_output_dir', help='Change where the mfcc output is going', default=None)
-    parser.add_argument('--segment_length', type=int, help='Length of each audio segment in milliseconds', default=5000)
+    parser.add_argument('--segment_length', type=int, help='Length of each audio segment in milliseconds', default=1000)
     parser.add_argument('--n_mfcc', type=int, help='Change the number of MFCC created', default=13)
     parser.add_argument('--set_brands', nargs='+', help='List of brands you want to include(instead of the top 10)', default=None)
+    parser.add_argument('--no_convert', help='Disable file conversion from .m4a to .wav', action='store_true', default=False)  
 
 
     args = parser.parse_args()
@@ -586,7 +599,7 @@ if __name__ == "__main__":
           f"Data Augmentation: {'Enabled' if args.augment else 'Disabled'}\n"
           f"Dataframe Creation: {'Enabled' if args.create_df else 'Disabled'}\n"
           f"Feature Extraction: {'Enabled' if args.extract_features else 'Disabled'}\n"
-          f"Mfcc Filename: {args.mfcc_filename}\n"
+          f"no_convert: {'Enabled' if args.no_convert else 'Disabled'}\n"
           f"N_MFCC: {args.n_mfcc}\n"
           f"Segment Length (ms): {args.segment_length}\n")
 
@@ -606,5 +619,6 @@ if __name__ == "__main__":
         mfcc_output_dir=args.mfcc_output_dir,
         n_mfcc=args.n_mfcc,
         segment_length_ms=args.segment_length,
-        set_brands=args.set_brands
+        set_brands=args.set_brands,
+        no_convert=args.no_convert
     )
